refactor: mogi_list response creation into a separate package
