+++ handler/mogi_button.go
@@ -83,7 +83,7 @@ func HandleMogiButtonClick(ctx context.Context, s *discordgo.Session, i *discord
 	}
 
 	// メッセージの作成
-	res, err := createMogiListInteractionResponse(guild.MogiList)
+	res, err := response.MakeMogiListInteractionResponse(guild.MogiList)
 	if err != nil {
 		s.FollowupMessageCreate(i.Interaction, true, response.MakeErrorWebhookParams(err))
 		return
